aider: Add support for 'application/msword' as a doc file.
# Context:
USER: add support for application/msword as a doc file
ASSISTANT: To add support for 'application/msword' as a doc file, we need to modify the 'app.py' file where the mimetype of the uploaded file is checked. We will add an additional condition to check if the mimetype is 'application/msword' and if so, assign the 'DocHandler' to the 'handler' variable.

Here is the change to be made in the 'app.py' file:

```python
app.py
<<<<<<< ORIGINAL
    if mimetype == 'application/pdf':
        handler = PDFHandler()
    elif mimetype == 'application/vnd.openxmlformats-officedocument.wordprocessingml.document':
        handler = DocHandler()
    elif mimetype == 'text/plain':
        handler = TextFileHandler()
=======
    if mimetype == 'application/pdf':
        handler = PDFHandler()
    elif mimetype in ['application/vnd.openxmlformats-officedocument.wordprocessingml.document', 'application/msword']:
        handler = DocHandler()
    elif mimetype == 'text/plain':
        handler = TextFileHandler()
>>>>>>> UPDATED
```

This change adds 'application/msword' to the list of mimetypes that are handled by the 'DocHandler'.